Fount the float

+++ train_controller/trainapi/commons/locomotive.py
@@ -109,7 +109,7 @@ def accelerate(locomotive_id, speed):
     if (motor.throttle == None):
         motor.throttle = 0
     
-    currentSpeed = motor.throttle*10
+    currentSpeed = int(motor.throttle*10)
     if (speed > 10 or speed < -10):
         raise Exception('Specified speed invalid. Speed values must be between 0 and 10')
     elif (currentSpeed > speed):
@@ -132,7 +132,7 @@ def decelerate(locomotive_id, speed):
     """
 
     motor = getattr(mh, "motor{}".format(locomotive_id))
-    currentSpeed = motor.throttle*10
+    currentSpeed = int(motor.throttle*10)
     if (speed > 10 or speed < -10):
         raise Exception('Specified speed invalid. Speed values must be between 0 and 10')
     elif (currentSpeed < speed):
